+++ lib/osm/due_badges.rb
@@ -35,11 +35,12 @@ module Osm
     # @!macro options_get
     # @return [Osm::DueBadges]
     def self.get(api, section, term=nil, options={})
+      require_ability_to(api, :read, :badge, section, options)
       section = Osm::Section.get(api, section, options) if section.is_a?(Fixnum)
       term_id = (term.nil? ? Osm::Term.get_current_term_for_section(api, section, options) : term).to_i
       cache_key = ['due_badges', section.id, term_id]
 
-      if !options[:no_cache] && cache_exist?(api, cache_key) && get_user_permission(api, section.id, :badge).include?(:read)
+      if !options[:no_cache] && cache_exist?(api, cache_key)
         return cache_read(api, cache_key)
       end