+++ lib/osm/register.rb
@@ -9,11 +9,12 @@ module Osm
     # @!macro options_get
     # @return [Array<Osm::Register::Field>] representing the fields of the register
     def self.get_structure(api, section, term=nil, options={})
+      Osm::Model.require_ability_to(api, :read, :register, section, options)
       section_id = section.to_i
       term_id = term.nil? ? Osm::Term.get_current_term_for_section(api, section).id : term.to_i
       cache_key = ['register_structure', section_id, term_id]
 
-      if !options[:no_cache] && Osm::Model.cache_exist?(api, cache_key) && Osm::Model.get_user_permission(api, section_id, :register).include?(:read)
+      if !options[:no_cache] && Osm::Model.cache_exist?(api, cache_key)
         return Osm::Model.cache_read(api, cache_key)
       end
 
@@ -45,11 +46,12 @@ module Osm
     # @!macro options_get
     # @return [Array<Register::Attendance>] representing the attendance of each member
     def self.get_attendance(api, section, term=nil, options={})
+      Osm::Model.require_ability_to(api, :read, :register, section, options)
       section_id = section.to_i
       term_id = term.nil? ? Osm::Term.get_current_term_for_section(api, section).id : term.to_i
       cache_key = ['register_attendance', section_id, term_id]
 
-      if !options[:no_cache] && Osm::Model.cache_exist?(api, cache_key) && Osm::Model.get_user_permission(api, section_id, :register).include?(:read)
+      if !options[:no_cache] && Osm::Model.cache_exist?(api, cache_key)
         return Osm::Model.cache_read(api, cache_key)
       end
 
@@ -94,8 +96,10 @@ module Osm
       raise ArgumentIsInvalid, ':section is missing' if data[:section].nil?
       raise ArgumentIsInvalid, ':evening is missing' if data[:evening].nil?
       raise ArgumentIsInvalid, ':members is missing' if data[:members].nil?
-
+      raise ArgumentIsInvalid, ':api is missing' if data[:api].nil?
       api = data[:api]
+      Osm::Model.require_ability_to(api, :write, :register, data[:section])
+
       term_id = data[:term].nil? ? Osm::Term.get_current_term_for_section(api, section).id : data[:term].to_i
 
       data[:members] = [*data[:members]].map{ |member| (member.is_a?(Fixnum) ? member : member.id).to_s } # Make sure it's an Array of Strings